Allow multiple admin nicks
This change just swaps the admin property on the bot config to be a vector
of strings instead of a single string. The Watcher value is now a
HashSet<String>, which is a little bit bigger change, but, because there
was already a function we use to determine if a command comes from an
admin or not, we only had to change that code in a single place.